% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliopts.r
\name{cli_process}
\alias{cli_process}
\title{Initialize the Command Line Interface options}
\usage{
cli_process(
  domain_opts,
  default_infile,
  default_outfile,
  default_x,
  default_y,
  default_groupby,
  default_facetvar
)
}
\arguments{
\item{domain_opts}{The list of potential command line arguments}

\item{default_infile}{Default value for the name of the dataset file to load}

\item{default_outfile}{Default value for the name of the output graph to export}

\item{default_x}{Default field to use for the x axis}

\item{default_y}{Default field to use for the y axis}

\item{default_groupby}{Default field to group the data into}

\item{default_facetvar}{Default field to segment the facets by}
}
\value{
The list of arguments provided on the cmd line by the user
}
\description{
This function processes the command line interface options entered by a user
and returns their input.
}
