% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.r
\name{export_barplot}
\alias{export_barplot}
\title{Exports a barplot}
\usage{
export_barplot(data, opt)
}
\arguments{
\item{data}{The dataset to use in the graph}
}
\value{
The original dataset (unmodified)
}
\description{
This function exports a dataset into a barplot based on user-specified settings. It assumes
that the xaxis, yaxis, and groupby are all fields within the dataset. The following fields
are required to be set as opts in opt: xaxis, yaxis, digits, groupby, outfile
}
