% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliopts.r
\name{finalize_opts}
\alias{finalize_opts}
\title{Parse the provided Command Line Interface arguments}
\usage{
finalize_opts(option_list)
}
\arguments{
\item{option_list}{The list of potential command line arguments}
}
\value{
The list of provided arguments
}
\description{
This function parses the command line interface arguments as values provided
by the init_opts function as well as any user-specified values
}
