% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliopts.r
\name{init_opts}
\alias{init_opts}
\title{Initialize the Command Line Interface options}
\usage{
init_opts(
  default_infile,
  default_outfile,
  default_x,
  default_y,
  default_groupby,
  default_facetvar
)
}
\arguments{
\item{default_infile}{Default value for the name of the dataset file to load}

\item{default_outfile}{Default value for the name of the output graph to export}

\item{default_x}{Default field to use for the x axis}

\item{default_y}{Default field to use for the y axis}

\item{default_groupby}{Default field to group the data into}

\item{default_facetvar}{Default field to facet the data by}
}
\value{
A list of options available
}
\description{
This function initializes a set of options that are used in generating
the graph.
}
