% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.r
\name{try_filter}
\alias{try_filter}
\title{Try to filter a column if a value is provided the the cmd line options}
\usage{
try_filter(., col, val)
}
\arguments{
\item{.}{The dataset}

\item{col}{The name of the column to filter}

\item{val}{The value to filter the column by}
}
\value{
The dataset where all values in col are val
}
\description{
This function filters a specified column within a dataset by a commnd line option.
}
